Implement core structure
This commit will implement the gist of states, transitions and operations. It will also make changes to the blueprint, where a decision was made to exclude valid transition states from the state class, and rather make the object graph take care of that. Note that this commit comments out build code for ROS in CMakeList and only make it compile the vanilla C++.