Make include folder include/fluid and update paths
This commit does a big refactor of the project where we make it
better structured for others use.
E.g. previously there was no namespace
when including files from fluid, it
would simply be #include <core/core.h>,
not #include <fluid/core/core.h>, which
it is change to now. This commit will also
remove the use of relative paths for header files.